projects: adalm-lsmspg: fix incompatible-pointer-types for GCC 14+
GCC 14 makes -Wincompatible-pointer-types an error. In the curvetrace
example declare temp_raw as uint16_t to match lm75_read_temperature(); it
feeds lm75_raw_to_millicelsius(uint16_t) and a %u print, both correct at
16-bit.

+++ projects/adalm-lsmspg/src/examples/curvetrace_example/curvetrace_example.c
@@ -719,7 +719,7 @@ int lm75_example(struct no_os_uart_desc *uart_desc)
 		.i2c_ip = &lm75_i2c_ip,
 	};
 	int ret;
-	uint32_t temp_raw;
+	uint16_t temp_raw;
 	uint32_t temp;
